snapventure-backend/requirements.txt
Django==1.10.3
djangorestframework==3.5.2
mysqlclient==1.3.9
Pillow==3.4.2
qrcode==5.3
six==1.10.0
